+++ src/cutty/projects/create.py
@@ -2,7 +2,10 @@
 from collections.abc import Callable
 from pathlib import Path
 
+import pygit2
+
 from cutty.repositories.domain.repository import Repository as Template
+from cutty.util import git
 
 
 LATEST_BRANCH = "cutty/latest"
@@ -10,3 +13,21 @@ UPDATE_BRANCH = "cutty/update"
 
 
 CreateProject = Callable[[Path], Template]
+
+
+def creategitrepository(projectdir: Path, template: Template) -> None:
+    """Create a git repository."""
+    try:
+        project = git.Repository.open(projectdir)
+    except pygit2.GitError:
+        project = git.Repository.init(projectdir)
+
+    project.commit(message=_commitmessage(template))
+    project.heads[LATEST_BRANCH] = project.head.commit
+
+
+def _commitmessage(template: Template) -> str:
+    if template.revision:
+        return f"Initial import from {template.name} {template.revision}"
+    else:
+        return f"Initial import from {template.name}"
